a2a/python-agent/requirements.txt
a2a-python>=0.2
uvicorn>=0.29
PyJWT>=2.9
cryptography>=43.0
httpx>=0.27
